+++ public/index.php
@@ -13,6 +13,7 @@ ini_set('display_errors', 1);
 
 use DLRoute\Requests\DLRoute;
 use DLTools\Compilers\DLView;
+use DLTools\Database\Model;
 
 $sessionExpirte = time() + 1300;