+++ src/Database/DLQueryBuilder.php
@@ -7,7 +7,7 @@ namespace DLTools\Database;
  * 
  * @package DLTools\Database
  * 
- * @version 1.0.0 (release)
+ * @version v0.1.63
  * @author David E Luna M <davidlunamontilla@gmail.com>
  * @copyright 2023 David E Luna M
  * @license MIT
@@ -200,21 +200,41 @@ trait DLQueryBuilder {
      * resultados según un conjunto de valores en un campo específico de la base de datos.
      *
      * Ejemplo de uso:
-     * ```php
-     * $queryBuilder->where_in('campo', 'valor1', 'valor2', 'valor3');
      * ```
-     * Generará una cláusula SQL similar a:
-     * ```sql
-     * WHERE campo IN ('valor1', 'valor2', 'valor3')
+     * <?php
+     * $queryBuilder->where_in('campo', ['valor1', 'valor2', 'valor3']);
      * ```
      *
-     * @param string $field El nombre del campo sobre el cual se aplicará la condición.
-     * @param string ...$value Uno o más valores para la cláusula "IN".
-     * @return static Retorna la instancia actual para permitir encadenamiento de métodos.
+     * Generará una cláusula SQL similar a:
+     * ```
+     * <?php
+     * WHERE campo IN (':in_campo1', ':in_campo2', ':in_campo3')
+     * ```
+     * 
+     * Donde `:in_campo1`, `:in_campo2` y `:in_campo3` es el marcador de posición de cada valor
+     *
+     * @param string   $field   El nombre del campo sobre el cual se aplicará la condición.
+     * @param string[] $values  Lista de valores para la cláusula "IN".
+     * @param string   $logical Operador lógico para combinar condiciones (por defecto, DLDatabase::AND).
+     * @return DLDatabase   Devuelve una instancia DLDatabase para permitir el encadenamiento.
+     *
+     * @since v0.1.63 Se actualizó la firma del método para aceptar un array de valores en lugar de parámetros individuales
+     *                y se agregó un tercer parámetro para definir el operador lógico.
      */
-    public function where_in(string $field, string ...$values): static {
+    public function where_in(string $field, array $values, string $logical = DLDatabase::AND): DLDatabase {
         $string_values = $this->get_string_values($field, ...$values);
-        $this->where = "WHERE {$field} IN ({$string_values})";
+
+        /** @var string $condition */
+        $condition = "{$field} IN ({$string_values})";
+
+        $logical = $this->get_logical_operator($logical);
+
+        if (count($this->conditions) > 0) {
+            $condition = "{$logical} {$condition}";
+        }
+
+        $this->conditions[] = $condition;
+        $this->where = "WHERE " . implode(" ", $this->conditions);
 
         return $this;
     }
@@ -234,7 +254,7 @@ trait DLQueryBuilder {
      * @param string ...$values Una lista de valores a formatear.
      * @return string Una cadena de texto con los valores formateados para su uso en consultas SQL.
      */
-    private function get_string_values(string $field = 'field', string ...$values): string {
+    protected function get_string_values(string $field = 'field', string ...$values): string {
         /** @var string[] $fragments */
         $fragments = [];
 
@@ -243,10 +263,135 @@ trait DLQueryBuilder {
 
         foreach ($values as $value) {
             ++$index;
-            array_push($fragments, ":{$field}{$index}");
-            $this->param[":{$field}{$index}"] = $value;
+
+            /** @var string $key */
+            $key = ":in_{$field}{$index}";
+            array_push($fragments, $key);
+
+            $this->param[$key] = $value;
         }
 
         return implode(", ", $fragments);
+    }
+
+    /**
+     * Construye la consulta condicional `where` estableciendo la condición.
+     *
+     * @param array $conditions Condicionales almacenadas.
+     * @param string $field Campo o columna de la tabla de referencia para ser consultada.
+     * @param string $operator Operador de comparación.
+     * @param string|null $value Valor de la consulta.
+     * @param string $logical Operador lógico.
+     * @return void
+     */
+    protected function set_conditions(array &$conditions, string $field, string $operator, ?string $value = NULL, string $logical = 'AND') {
+        $operator = strtoupper(trim($operator));
+
+        /** @var string | null $key */
+        $key = $this->get_param_key($field, $operator, $value);
+
+        /** @var string $condition */
+        $condition = "";
+
+        if (is_null($value) && !is_null($key)) {
+            $operator = "=";
+        }
+
+        $condition = $key
+            ? "{$field} {$operator} {$key}"
+            : "{$field} {$operator}";
+
+        if (count($conditions) > 0) {
+            $condition = "{$logical} {$condition}";
+        }
+
+        $conditions[] = $condition;
+    }
+
+    /**
+     * Devuelve el operador lógico
+     *
+     * @param string $logical Operador lógico a ser procesado
+     * @return string
+     */
+    protected function get_logical_operator(string $logical): string {
+        $logical = strtoupper(trim($logical));
+
+        /** @var string[] $allowed */
+        $allowed = ["AND", "OR"];
+
+        return in_array($logical, $allowed) ? $logical : 'AND';
+    }
+
+    /**
+     * Establece y devuelve el parámetro con el valor asociado a una clave de tipo :key.
+     *
+     * @param string $field Campo o columna de la tabla.
+     * @param string $operator Operator o valor según sea el caso.
+     * @param string|null $value Valor de la consulta.
+     * @return string|null
+     */
+    protected function get_param_key(string $field, string $operator, ?string $value): ?string {
+
+        if ($this->is_null_operator($operator)) {
+            return null;
+        }
+
+        static $counters = [];
+
+        if (!isset($counters[$field])) {
+            $counters[$field] = 0;
+        }
+
+        ++$counters[$field];
+
+        /** @var string $key */
+        $key = ":{$field}" . $counters[$field];
+
+        $this->set_param_key($key, $operator, $value);
+        return $key;
+    }
+
+
+    /**
+     * Establece la clave del parámetro de la consulta parametrizada.
+     *
+     * @param string $key Clave del parámetro
+     * @param string $operator Operador
+     * @param string|null $value Valor que se utilizar para ser consultado
+     * @return void
+     */
+    protected function set_param_key(string $key, string $operator, ?string $value): void {
+        $this->param[$key] = !is_null($value)
+            ? trim($value)
+            : trim($operator);
+    }
+
+    /**
+     * Devuelve el operador normalizado.
+     *
+     * @param string $operator Operador a ser procesado.
+     * @return string Operador en su forma estándar.
+     */
+    protected function get_operator(string $operator): string {
+        $operator = strtoupper(trim($operator));
+
+        return match ($operator) {
+            "IS NULL", "IS NOT NULL" => $operator,
+            default => $operator
+        };
+    }
+
+    /**
+     * Valida si se trata de un operador de nulidad
+     *
+     * @param string $operator Operador a verificar
+     * @return boolean
+     */
+    protected function is_null_operator(string $operator): bool {
+        /** @var string[] $null_operators */
+        $null_operators = ["IS NULL", "IS NOT NULL"];
+
+        return in_array($operator, $null_operators, true);
     }
 }